requirements.txt
# Essentials
platformdirs~=4.2.2
requests~=2.32.3
websocket-client~=1.8.0
# Only essential if you don't have it installed in your system package manager.
PySide6~=6.7.1


# Depedencies of dependencies. You need not install these manually, but here for good measure.
certifi~=2024.6.2
charset-normalizer~=3.3.2
idna~=3.7
PySide6_Addons~=6.7.1
PySide6_Essentials~=6.7.1
shiboken6~=6.7.1
urllib3~=2.2.1
